rag_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.3
python-multipart==0.0.12
pypdf==5.1.0
qdrant-client==1.12.1
numpy==2.1.3
openai==1.58.1
tenacity==9.0.0
python-dotenv==1.0.1
